ultima entrega

+++ JS/carrito.js
@@ -33,6 +33,7 @@ function cargarProductosCarrito(){
             </div>
             <div class="carrito-producto-cantidad">
                 <small>Cantidad</small>
+                <input type="number" value="1" class="cart-quantity">
                 <p>${producto.cantidad}</p>
             </div>
             <div class="carrito-producto-precio">
@@ -70,6 +71,25 @@ function actualizarBotonesEliminar() {
     });
 }
 function eliminarDelCarrito(e) {
+    Toastify({
+        text: "Producto Eliminado",
+        duration: 3000,
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background: "linear-gradient(to right, #552121, #9c5252)",
+          borderRadius: "2rem",
+          textTransform: "uppercase",
+          fontSize: ".75rem"
+        },
+        offset: {
+            x: '1.5rem', // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+            y: '1.5rem' // vertical axis - can be a number or a string indicating unity. eg: '2em'
+          },
+        onClick: function(){} // Callback after click
+      }).showToast();
     let idBoton = e.currentTarget.id;
     const index = productosEnCarrito.findIndex(producto => producto.id === idBoton);
     productosEnCarrito.splice(index, 1);
@@ -80,9 +100,21 @@ function eliminarDelCarrito(e) {
 botonVaciar.addEventListener("click", vaciarCarrito)
 
 function vaciarCarrito() {
-    productosEnCarrito.length = 0
-    localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito))
-    cargarProductosCarrito();
+    Swal.fire({
+        title: '¿Estás seguro?',
+        icon: 'question',
+        html: `Se van a borrar ${productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos.`,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: 'Sí',
+        cancelButtonText: 'No'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+            cargarProductosCarrito();
+        }
+      })
 
 }
